engine/requirements.txt
fastapi
uvicorn
slither-analyzer
google-generativeai
pydantic
python-dotenv
httpx
pytest
pytest-asyncio